Restructure engine.rs
